Update main.py
Really interesting here. The behavior is a spiral from the center that tends to be toward the upper half. 
Reducing the time the robot can drive straight to 1.2 seconds is what keeps the robot finding trash without having to go too far. It is
enough to stay in the area defined by that circle to stay in the range of 50+ kg.